Reflect feedback and a fix
* Remove unused constants
* Use BindingFlags.DoNotWrapExceptions
* Remove checking exception messages
* Fix coding convention w.r.t. var keyword usage
* Remove the _unloadCount field that was a remainder of a previous state
of the code
* Disable COM and delegate marshalling tests for Unix
