+++ src/System.Runtime.Loader/tests/Configurations.props
@@ -2,7 +2,8 @@
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <BuildConfigurations>
-      netcoreapp;
+      netcoreapp-Unix;
+      netcoreapp-Windows_NT;
       uap;
     </BuildConfigurations>
   </PropertyGroup>
