+++ src/System.Runtime.Loader/tests/System.Runtime.Loader.Tests.csproj
@@ -8,6 +8,7 @@
   <ItemGroup>
     <Compile Include="AssemblyLoadContextTest.cs" />
     <Compile Include="CollectibleAssemblyLoadContextTest.cs" />
+    <Compile Include="CollectibleAssemblyLoadContextTest.Windows.cs" Condition="'$(TargetsWindows)' == 'true'"/>
     <Compile Include="CustomTPALoadContext.cs" />
     <Compile Include="ResourceAssemblyLoadContext.cs" />
   </ItemGroup>
